requirements.txt
bleak==0.21.1
cfgv==3.4.0
dbus-fast==2.21.0
distlib==0.3.8
filelock==3.13.1
identify==2.5.33
nodeenv==1.8.0
Pillow==10.1.0
platformdirs==4.1.0
plyer==2.1.0
pre-commit==3.6.0
pystray==0.19.5
python-xlib==0.33
PyYAML==6.0.1
setuptools==68.2.2
six==1.16.0
virtualenv==20.25.0
wheel==0.41.2
